Upload assetlinks.json to be hosted at https://<domain>/.well-known/assetlinks.json
This file is not meant to be committed to git, but it is included here as an example.

What is this file for?

This is used to verify that the app is allowed to open links to your domain. This is required for deep linking to work on Android.

The sha256 must be retrieved from the key used to sign the app

To verify that the sha256 is correct, you can use the following command
adb shell pm get-app-links com.example.flutterbasicssamples (your app package name)